Selenium java browser test updated: added the following
added support for tests using Opera and firefox browsers to...
WebDriverCheckboxTest.java
WebDriverRadioButtons.java
WebDriverSelectItem.java

Chrome, Firefox, and Opera can now use this test as long as the commas are removed and replaced... and example

        //WebDriver driver = new FirefoxDriver();
        WebDriver driver = new ChromeDriver();
        //WebDriver driver = new OperaDriver();

this displays the use of the ChromeDriver over the other two because the other two are commented out. to use them replace and remove comments marks as needed.